WIP Consolidated code for crypto rules; need to update messages
