+++ src/Microsoft.NetCore.Analyzers/Core/Security/Helpers/SecurityMemberNames.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace Microsoft.NetCore.Analyzers.Security.Helpers
+{
+    internal static class SecurityMemberNames
+    {
+        public const string CreateSignature = "CreateSignature";
+    }
+}
