+++ src/Microsoft.NetCore.Analyzers/Core/Security/Helpers/SecurityTypes.cs
@@ -10,13 +10,54 @@ namespace Microsoft.NetCore.Analyzers.Security.Helpers
         {
             return compilation.GetTypeByMetadataName("System.Security.Cryptography.MD5");
         }
+
         public static INamedTypeSymbol SHA1(Compilation compilation)
         {
             return compilation.GetTypeByMetadataName("System.Security.Cryptography.SHA1");
         }
+
         public static INamedTypeSymbol HMACSHA1(Compilation compilation)
         {
             return compilation.GetTypeByMetadataName("System.Security.Cryptography.HMACSHA1");
+        }
+        public static INamedTypeSymbol DES(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.DES");
+        }
+
+        public static INamedTypeSymbol DSA(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.DSA");
+        }
+
+        public static INamedTypeSymbol DSASignatureFormatter(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.DSASignatureFormatter");
+        }
+
+        public static INamedTypeSymbol HMACMD5(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.HMACMD5");
+        }
+
+        public static INamedTypeSymbol RC2(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.RC2");
+        }
+
+        public static INamedTypeSymbol TripleDES(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.TripleDES");
+        }
+
+        public static INamedTypeSymbol RIPEMD160(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.RIPEMD160");
+        }
+
+        public static INamedTypeSymbol HMACRIPEMD160(Compilation compilation)
+        {
+            return compilation.GetTypeByMetadataName("System.Security.Cryptography.HMACRIPEMD160");
         }
     }
 }
