+++ src/Microsoft.NetFramework.Analyzers/Core/Helpers/CompilationSecurityTypes.cs
@@ -26,14 +26,7 @@ namespace Microsoft.NetFramework.Analyzers.Helpers
         public INamedTypeSymbol XmlResolver { get; private set; }
         public INamedTypeSymbol XmlSecureResolver { get; private set; }
         public INamedTypeSymbol XsltSettings { get; private set; }
-        public INamedTypeSymbol DES { get; private set; }
-        public INamedTypeSymbol DSA { get; private set; }
-        public INamedTypeSymbol DSASignatureFormatter { get; private set; }
-        public INamedTypeSymbol HMACMD5 { get; private set; }
-        public INamedTypeSymbol RC2 { get; private set; }
-        public INamedTypeSymbol TripleDES { get; private set; }
-        public INamedTypeSymbol RIPEMD160 { get; private set; }
-        public INamedTypeSymbol HMACRIPEMD160 { get; private set; }
+
 
         public CompilationSecurityTypes(Compilation compilation)
         {
@@ -42,14 +35,6 @@ namespace Microsoft.NetFramework.Analyzers.Helpers
             SystemObject = SecurityTypes.SystemObject(compilation);
             SystemException = SecurityTypes.SystemException(compilation);
             SystemSystemException = SecurityTypes.SystemSystemException(compilation);
-            DES = SecurityTypes.DES(compilation);
-            DSA = SecurityTypes.DSA(compilation);
-            DSASignatureFormatter = SecurityTypes.DSASignatureFormatter(compilation);
-            HMACMD5 = SecurityTypes.HMACMD5(compilation);
-            RC2 = SecurityTypes.RC2(compilation);
-            TripleDES = SecurityTypes.TripleDES(compilation);
-            RIPEMD160 = SecurityTypes.RIPEMD160(compilation);
-            HMACRIPEMD160 = SecurityTypes.HMACRIPEMD160(compilation);
             XmlDocument = SecurityTypes.XmlDocument(compilation);
             XPathDocument = SecurityTypes.XPathDocument(compilation);
             XmlSchema = SecurityTypes.XmlSchema(compilation);
